+++ src/vue/modules/requests/incoming-withdrawal-requests/components/request-viewer.vue
@@ -1,9 +1,14 @@
 <template>
   <div class="request-viewer">
+    <asset-summary-viewer
+      :asset-code="request.asset.code"
+      :asset-name="request.asset.name"
+      :asset-logo-url="assetLogoUrl"
+    />
+
     <request-message-viewer
       class="request-viewer__state-message"
       :request="request"
-      direction="incoming"
     />
 
     <request-attributes-viewer
@@ -20,11 +25,15 @@
 </template>
 
 <script>
+import AssetSummaryViewer from '../../shared/components/asset-summary-viewer'
+
 import RequestMessageViewer from './request-message-viewer'
 import RequestAttributesViewer from './request-attributes-viewer'
 import RequestActions from './request-actions'
 
 import { IncomingWithdrawalRequest } from '../wrappers/incoming-withdrawal-request'
+
+import config from '@/config'
 
 const EVENTS = {
   requestUpdated: 'request-updated',
@@ -33,6 +42,7 @@ const EVENTS = {
 export default {
   name: 'request-viewer',
   components: {
+    AssetSummaryViewer,
     RequestMessageViewer,
     RequestAttributesViewer,
     RequestActions,
@@ -45,10 +55,24 @@ export default {
   data: _ => ({
     EVENTS,
   }),
+
+  computed: {
+    assetLogoUrl () {
+      if (this.request.asset) {
+        return this.request.asset.logoUrl(config.FILE_STORAGE)
+      } else {
+        return ''
+      }
+    },
+  },
 }
 </script>
 
 <style lang="scss" scoped>
+.request-viewer__state-message {
+  margin-top: 2rem;
+}
+
 .request-viewer__table {
   margin-top: 2rem;
 }