+++ src/vue/modules/requests/incoming-withdrawal-requests/components/requests-table.vue
@@ -11,6 +11,9 @@
             <th :title="'incoming-withdrawal-requests.requestor-header' | globalize">
               {{ 'incoming-withdrawal-requests.requestor-header' | globalize }}
             </th>
+            <th :title="'incoming-withdrawal-requests.asset-code-header' | globalize">
+              {{ 'incoming-withdrawal-requests.asset-code-header' | globalize }}
+            </th>
             <th :title="'incoming-withdrawal-requests.amount-header' | globalize">
               {{ 'incoming-withdrawal-requests.amount-header' | globalize }}
             </th>
@@ -31,6 +34,10 @@
           >
             <td :title="request.requestor">
               <email-getter :account-id="request.requestor" />
+            </td>
+
+            <td :title="request.asset.code">
+              {{ request.asset.code }}
             </td>
 
             <td :title="request.amount | formatMoney">